feat: 增加wget等依赖包；升级备份方法

+++ main.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
-if [ $# -eq 1 ]; then
-  version=$1
-else
-  version="$(curl -fsSLI -o /dev/null -w "%{url_effective}" https://github.com/sixmillions/memos-replit/releases/latest)"
-  version="${version#https://github.com/sixmillions/memos-replit/releases/tag/}"
-fi
-
 echo "down load memos ${version} >>> $(date '+%Y-%m-%d %H:%M:%S')"
 
-rm -rf memos && wget https://github.com/sixmillions/memos-replit/releases/download/${version}/memos
+if [ ! -e memos ];then
+    if [ $# -eq 1 ]; then
+      version=$1
+    else
+      version="$(curl -fsSLI -o /dev/null -w "%{url_effective}" https://github.com/sixmillions/memos-replit/releases/latest)"
+      version="${version#https://github.com/sixmillions/memos-replit/releases/tag/}"
+    fi
+    wget https://github.com/sixmillions/memos-replit/releases/download/${version}/memos
+fi
 
 echo "run memos >>> $(date '+%Y-%m-%d %H:%M:%S')"
